add a constants file to productdisplayitem.main.tsx
