+++ src/components/src/ProductDisplayItem/productdisplayitem.main.constants.ts
@@ -1,0 +1,6 @@
+export const REQUISITION_LISTS_ZOOM : [] = [
+    'itemlistinfo',
+    'itemlistinfo:allitemlists',
+    'itemlistinfo:allitemlists:element',
+    'itemlistinfo:allitemlists:element:additemstoitemlistform',
+  ];
